MA-6446: audio hal: change the threshold for times of writing fail
After HDMI plugout, there are a lot of print log from kernel which
may block the process in the user space. The log is caused by reopen
HDMI audio device, everytime open the HDMI device, there are about 120
character printed, which cost 120*64us= 7.68ms, which is too long.

The threshold for error number is 20, after this threshold, the hal will
try to open the HDMI device. so enlarge the threshold, then the times of
open will be reduced.

Signed-off-by: Shengjiu Wang <shengjiu.wang@freescale.com>